refactor: move AWS stuff into more logical spots
