+++ src/main/java/com/conveyal/datatools/common/utils/aws/CheckedAWSException.java
@@ -1,4 +1,4 @@
-package com.conveyal.datatools.common.utils;
+package com.conveyal.datatools.common.utils.aws;
 
 import com.amazonaws.AmazonServiceException;
